separated the Confirm Email into a separate controller
